refactor: move interchain tests to subfolder
