+++ e2e/interchaintests/ibc_to_hyperlane_test.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package e2e
+package interchaintests
 
 import (
 	"strings"
