+++ e2e/interchaintests/utils_queries.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package e2e
+package interchaintests
 
 import (
 	"bytes"
